refactored weatherDetails cells
