+++ weatherApp/Features/WeatherDetails/UI/Cell/SingleWeatherInformationCollectionViewCell.swift
@@ -8,29 +8,33 @@
 
 import UIKit
 import Kingfisher
-import PureLayout
 
 class SingleWeatherInformationCollectionViewCell: UICollectionViewCell {
     
-    private let headerLabel = UILabel(forAutoLayout: ())
-    private let mainInformationLabel = UILabel(forAutoLayout: ())
-    private let translucentView = UIView(forAutoLayout: ())
-    private let weatherIcon = UIImageView(forAutoLayout: ())
+    //MARK: - Properties
     
-    static let height: CGFloat = 140
     static var typeName: String {
         String(describing: self)
     }
     
+    var headerLabel: UILabel!
+    var mainInformationLabel: UILabel!
+    var translucentView: UIView!
+    var weatherIcon: UIImageView!
+    
+    //MARK: - Initialization
+    
     required init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
-        commonInit()
+        buildViews()
     }
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        commonInit()
+        buildViews()
     }
+    
+    //MARK: - Set data
     
     override func prepareForReuse() {
         super.prepareForReuse()
@@ -48,69 +52,5 @@ class SingleWeatherInformationCollectionViewCell: UICollectionViewCell {
             weatherIcon.kf.setImage(with: url)
         }
     }
-    
-    //MARK: - Styling UI Elements
-    
-    private func commonInit() {
-        setupSubviews()
-        styleHeaderLabel()
-        styleMainInformationLabel()
-        styleTranslucentView()
-        setupLayout()
-    }
-    
-    private func styleHeaderLabel() {
-        headerLabel.font = UIFont.systemFont(ofSize: 14, weight: .semibold)
-        headerLabel.textColor = .white
-        headerLabel.textAlignment = .center
-    }
-    
-    private func styleMainInformationLabel() {
-        mainInformationLabel.font = UIFont.systemFont(ofSize: 14, weight: .semibold)
-        mainInformationLabel.adjustsFontSizeToFitWidth = true
-        mainInformationLabel.minimumScaleFactor = 0.5
-        mainInformationLabel.textColor = .white
-        mainInformationLabel.numberOfLines = 0
-        mainInformationLabel.textAlignment = .center
-    }
-    
-    private func styleTranslucentView() {
-        translucentView.backgroundColor = .black20
-        translucentView.layer.cornerRadius = 15
-    }
-    
-    //MARK: - Setting Up Layout
-    
-    private func setupSubviews() {
-        addSubview(headerLabel)
-        addSubview(translucentView)
-        addSubview(mainInformationLabel)
-        addSubview(weatherIcon)
-    }
-    
-    private func setupLayout() {
-        let offset: CGFloat = 5
-        
-        heightAnchor.constraint(equalToConstant: SingleWeatherInformationCollectionViewCell.height).isActive = true
-        
-        headerLabel.autoPinEdge(.top, to: .top, of: contentView)
-        headerLabel.autoPinEdge(.trailing, to: .trailing, of: contentView)
-        headerLabel.autoPinEdge(.leading, to: .leading, of: contentView)
-        
-        translucentView.autoPinEdge(.top, to: .bottom, of: headerLabel, withOffset: offset)
-        translucentView.autoPinEdge(.trailing, to: .trailing, of: contentView)
-        translucentView.autoPinEdge(.leading, to: .leading, of: contentView)
-        translucentView.autoPinEdge(.bottom, to: .bottom, of: contentView, withOffset: -offset)
-        
-        weatherIcon.autoSetDimension(.height, toSize: 50, relation: .lessThanOrEqual)
-        weatherIcon.autoSetDimension(.width, toSize: 50, relation: .lessThanOrEqual)
-        weatherIcon.autoPinEdge(.top, to: .top, of: translucentView, withOffset: offset)
-        weatherIcon.autoAlignAxis(.vertical, toSameAxisOf: translucentView)
-        
-        mainInformationLabel.autoPinEdge(.top, to: .bottom, of: weatherIcon)
-        mainInformationLabel.autoPinEdge(.bottom, to: .bottom, of: translucentView, withOffset: -offset)
-        mainInformationLabel.autoPinEdge(.leading, to: .leading, of: translucentView)
-        mainInformationLabel.autoPinEdge(.trailing, to: .trailing, of: translucentView)
-    }
-    
+
 }
